register

+++ 27-FrontToBackSqlConnection/Areas/AdminPanel/Controllers/ProductController.cs
@@ -1,4 +1,5 @@
 ﻿using _27_FrontToBackSqlConnection.Areas.AdminPanel.ViewModels;
+using _27_FrontToBackSqlConnection.Areas.AdminPanel.ViewModels.Products;
 using _27_FrontToBackSqlConnection.Data;
 using _27_FrontToBackSqlConnection.Models;
 using _27_FrontToBackSqlConnection.Utilities.Enums;
@@ -43,7 +44,8 @@ namespace _27_FrontToBackSqlConnection.Areas.AdminPanel.Controllers
         {
             ProductCreateVM productCreateVM = new()
             {
-                Categories = await _context.Categories.Where(c => !c.IsDeleted).ToListAsync()
+                Categories = await _context.Categories.Where(c => !c.IsDeleted).ToListAsync(),
+                Tags = await _context.Tags.Where(c => !c.IsDeleted).ToListAsync()
             };
             return View(productCreateVM);
         }
@@ -61,17 +63,17 @@ namespace _27_FrontToBackSqlConnection.Areas.AdminPanel.Controllers
                 ModelState.AddModelError(nameof(productCreateVM.MainPhoto), "File type is incorrect");
                 return View(productCreateVM);
             }
+            if (!productCreateVM.MainPhoto.CheckFileSize(FileSize.MB, 1))
+            {
+                ModelState.AddModelError(nameof(productCreateVM.MainPhoto), "File size  less be than 2mb");
+                return View(productCreateVM);
+            }
             if (!productCreateVM.HoverPhoto.CheckFileType("image/"))
             {
                 ModelState.AddModelError(nameof(productCreateVM.HoverPhoto), "File type is incorrect");
                 return View(productCreateVM);
             }
 
-            if (!productCreateVM.MainPhoto.CheckFileSize(FileSize.MB, 1))
-            {
-                ModelState.AddModelError(nameof(productCreateVM.MainPhoto), "File size  less be than 2mb");
-                return View(productCreateVM);
-            }
             if (!productCreateVM.HoverPhoto.CheckFileSize(FileSize.MB, 1))
             {
                 ModelState.AddModelError(nameof(productCreateVM.HoverPhoto), "File size  less be than 2mb");
@@ -122,21 +124,64 @@ namespace _27_FrontToBackSqlConnection.Areas.AdminPanel.Controllers
                 product.ProductTags = productCreateVM.TagIds.Select(tId => new ProductTag { TagId = tId }).ToList();
             }
 
+            string info = string.Empty;
+            if (productCreateVM.AdditionalPhotos is not null)
+            {
+                foreach (var file in productCreateVM.AdditionalPhotos)
+                {
+                    if (!file.CheckFileType("image/"))
+                    {
+                        info += $"<p>{file.FileName} type was not correct</p>";
+                        continue;
+                    }
+                    if (!file.CheckFileSize(FileSize.KB, 100))
+                    {
+                        info += $"<p>{file.FileName} size was not correct</p>";
+                        continue;
+                    }
+
+                    product.ProductImages.Add(new ProductImage
+                    {
+                        Image = await file.CreateFile(_env.WebRootPath, "assets", "images", "website-images"),
+                        IsPrimary = null
+                    });
+                }
+            }
+            TempData["FileInfo"] = info;
+
             await _context.Products.AddAsync(product);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
+
+
         public async Task<IActionResult> Update(int? id)
         {
-            if (id is null || id < 1) return BadRequest();
+            if (id == null || id < 1) return BadRequest();
+            Product? existProduct =
+                await _context.Products
+                    .Include(p => p.ProductImages)
+                    .Include(p => p.ProductTags)
+                    .FirstOrDefaultAsync(p => p.Id == id);
 
-            Category existCategory = await _context.Categories
-                .Where(c => !c.IsDeleted)
-                .FirstOrDefaultAsync(c => c.Id == id);
-            if (existCategory is null) return NotFound();
+            if (existProduct == null) return NotFound();
 
-            return View(existCategory);
+            if (!ModelState.IsValid) return View();
+            ProductUpdateVM productUpdateVM = new()
+            {
+                Name = existProduct.Name,
+                Price = (decimal)existProduct.Price,
+                Description = existProduct.Description,
+                SKU = existProduct.SKU,
+                CategoryId = existProduct.CategoryId,
+                TagIds = existProduct.ProductTags.Select(pt => pt.TagId).ToList(),
+                Categories = await _context.Categories.ToListAsync(),
+                Tags = await _context.Tags.ToListAsync(),
+                ProductImages = existProduct.ProductImages
+            };
+
+            return View(productUpdateVM);
         }
    
 
@@ -145,13 +190,48 @@ namespace _27_FrontToBackSqlConnection.Areas.AdminPanel.Controllers
         {
             if (id == null || id < 1) return BadRequest();
 
+            Product? existProduct = await _context.Products
+                .Include(p => p.ProductImages)
+                .Include(p => p.ProductTags)
+                .FirstOrDefaultAsync(p => p.Id == id);
+
+            if (existProduct == null) return NotFound();
+
             productUpdateVM.Categories = await _context.Categories.Where(c => !c.IsDeleted).ToListAsync();
             productUpdateVM.Tags = await _context.Tags.Where(t => !t.IsDeleted).ToListAsync();
+            productUpdateVM.ProductImages = existProduct.ProductImages;
 
             if (!ModelState.IsValid) return View(productUpdateVM);
 
-            Product? existProduct = await _context.Products.FirstOrDefaultAsync(p => p.Id == id);
-            if (existProduct == null) return NotFound();
+
+            if (productUpdateVM.MainPhoto is not null)
+            {
+                if (!productUpdateVM.MainPhoto.CheckFileType("image/"))
+                {
+                    ModelState.AddModelError(nameof(productUpdateVM.MainPhoto), "File type is incorrect");
+                    return View(productUpdateVM);
+                }
+                if (!productUpdateVM.MainPhoto.CheckFileSize(FileSize.MB, 1))
+                {
+                    ModelState.AddModelError(nameof(productUpdateVM.MainPhoto), "File size  less be than 2mb");
+                    return View(productUpdateVM);
+                }
+            }
+
+            if (productUpdateVM.HoverPhoto is not null)
+            {
+                if (!productUpdateVM.HoverPhoto.CheckFileType("image/"))
+                {
+                    ModelState.AddModelError(nameof(productUpdateVM.HoverPhoto), "File type is incorrect");
+                    return View(productUpdateVM);
+                }
+
+                if (!productUpdateVM.HoverPhoto.CheckFileSize(FileSize.MB, 1))
+                {
+                    ModelState.AddModelError(nameof(productUpdateVM.HoverPhoto), "File size  less be than 2mb");
+                    return View(productUpdateVM);
+                }
+            }
 
             bool existCategory = productUpdateVM.Categories.Any(c => c.Id == productUpdateVM.CategoryId);
             if (!existCategory)
@@ -175,18 +255,86 @@ namespace _27_FrontToBackSqlConnection.Areas.AdminPanel.Controllers
                 productUpdateVM.TagIds = new();
             }
 
-            _context.ProductTags.RemoveRange(existProduct.ProductTags
-                .Where(pTag => !productUpdateVM.TagIds
-                    .Exists(tId => tId == pTag.TagId))
-                .ToList());
+            if (productUpdateVM.TagIds is not null)
+            {
+                _context.ProductTags.RemoveRange(existProduct.ProductTags
+                    .Where(pTag => !productUpdateVM.TagIds
+                        .Exists(tId => tId == pTag.TagId))
+                    .ToList());
 
-            _context.ProductTags.AddRange(productUpdateVM.TagIds
-                .Where(tId => !existProduct.ProductTags.Exists(pTag => pTag.TagId == tId))
-                .ToList()
-                .Select(tId => new ProductTag { TagId = tId, ProductId = existProduct.Id }));
+                _context.ProductTags.AddRange(productUpdateVM.TagIds
+                    .Where(tId => !existProduct.ProductTags.Exists(pTag => pTag.TagId == tId))
+                    .ToList()
+                    .Select(tId => new ProductTag { TagId = tId, ProductId = existProduct.Id }));
+            }
+
+            if (productUpdateVM.MainPhoto is not null)
+            {
+                string fileName = await productUpdateVM.MainPhoto.CreateFile(_env.WebRootPath, "assets", "images", "website-images");
+                ProductImage mainImage = existProduct.ProductImages.FirstOrDefault(p => p.IsPrimary == true);
+
+                mainImage.Image.DeleteFile(_env.WebRootPath, "assets", "images", "website-images");
+                existProduct.ProductImages.Remove(mainImage);
+                existProduct.ProductImages.Add(new ProductImage
+                {
+                    Image = fileName,
+                    IsPrimary = true
+                });
+            }
+            if (productUpdateVM.HoverPhoto is not null)
+            {
+                string fileName = await productUpdateVM.HoverPhoto.CreateFile(_env.WebRootPath, "assets", "images", "website-images");
+                ProductImage hoverImage = existProduct.ProductImages.FirstOrDefault(p => p.IsPrimary == false);
+
+                hoverImage.Image.DeleteFile(_env.WebRootPath, "assets", "images", "website-images");
+                existProduct.ProductImages.Remove(hoverImage);
+                existProduct.ProductImages.Add(new ProductImage
+                {
+                    Image = fileName,
+                    IsPrimary = false
+                });
+            }
+
+            if (productUpdateVM.ImageIds is not null)
+            {
+                productUpdateVM.ImageIds = new List<int>();
+            }
+
+            var deleteImages =
+                existProduct.ProductImages
+                    .Where(pi => productUpdateVM.ImageIds
+                        .Exists(imgId => imgId == pi.Id) && pi.IsPrimary == true)
+                    .ToList();
+            deleteImages.ForEach(di => di.Image.DeleteFile(_env.WebRootPath, "assets", "images", "website-images"));
+            _context.ProductImages.RemoveRange(deleteImages);
+
+            if (productUpdateVM.AdditionalPhotos is not null)
+            {
+                string info = string.Empty;
+                foreach (var file in productUpdateVM.AdditionalPhotos)
+                {
+                    if (!file.CheckFileType("image/"))
+                    {
+                        info += $"<p>{file.FileName} type was not correct</p>";
+                        continue;
+                    }
+                    if (!file.CheckFileSize(FileSize.KB, 100))
+                    {
+                        info += $"<p>{file.FileName} size was not correct</p>";
+                        continue;
+                    }
+
+                    existProduct.ProductImages.Add(new ProductImage
+                    {
+                        Image = await file.CreateFile(_env.WebRootPath, "assets", "images", "website-images"),
+                        IsPrimary = null
+                    });
+                }
+                TempData["FileInfo"] = info;
+            }
 
             existProduct.Name = productUpdateVM.Name;
-            existProduct.Price = productUpdateVM.Price;
+            existProduct.Price = (double)productUpdateVM.Price;
             existProduct.Description = productUpdateVM.Description;
             existProduct.SKU = productUpdateVM.SKU;
             existProduct.CategoryId = productUpdateVM.CategoryId.Value;
